+++ src/cosmicrays/WMAP07CRDensity.cpp
@@ -26,7 +26,7 @@ void WMAP07CRDensity::setParameters() {
 
 void WMAP07CRDensity::makeEnergyRange() {
 	QEnergy energy = minE;
-	float energyRatio =
+	double energyRatio =
 		exp(1./static_cast<double>(steps-1) * log(maxE/minE));
 
 	for (int i = 0; i < steps; ++i) {
