+++ test/testPiZeroIntegrator.cpp
@@ -140,7 +140,7 @@ TEST(PiZeroIntegrator, BremsstrahlungLOS) {
         std::chrono::time_point<std::chrono::system_clock> stop = std::chrono::system_clock::now();
 
         auto milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(stop - start);
-	float pxl_speed = milliseconds.count()/skymap->getNpix()*getThreadsNumber();
+	unsigned long pxl_speed = milliseconds.count()/skymap->getNpix()*getThreadsNumber();
 
 	std::cerr << "pxl spd: " << pxl_speed << " ms" << std::endl;
 	//skymap->save(output);
@@ -166,7 +166,7 @@ TEST(PiZeroIntegrator, PerformanceTest) {
         std::chrono::time_point<std::chrono::system_clock> stop = std::chrono::system_clock::now();
 
         auto milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(stop - start);
-	float pxl_speed = milliseconds.count()/skymap->getNpix()*getThreadsNumber();
+	unsigned long pxl_speed = milliseconds.count()/skymap->getNpix()*getThreadsNumber();
 
 	std::cerr << "pxl spd: " << pxl_speed << " ms" << std::endl;
