+++ test/testSynchroIntegrator.cpp
@@ -132,7 +132,7 @@ TEST(SynchroIntegrator, PerformanceTest) {
         std::chrono::time_point<std::chrono::system_clock> stop = std::chrono::system_clock::now();
 
         auto milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(stop - start);
-	float pxl_speed = milliseconds.count()/skymap->getNpix()*getThreadsNumber();
+	unsigned long pxl_speed = milliseconds.count()/skymap->getNpix()*getThreadsNumber();
 
         EXPECT_LE(pxl_speed, 200); // ms
 }
